Update git_cwd_info to indicate a detached head
If @git_parse_branch is still nil after trying to parse the branch,
set the color to red and print DETACHED as a reminder that I'm not
on a branch.

Also cleaned up the color assignment to happen in the print statement
instead of the helper methods, with the exception of when we're
detached.