git: Use non-work email for git config
On the off-chance that I were to leave thoughtbot, it would really suck
to lose all of my contributions on GitHub. Instead of attaching my
identity to a possibly temporary email address, lets attach it to one I
control.

Also, I've had this domain _for fucking ever_ and this seems like the
perfect use case for it.