zsh: Source zsh config first and reorganize
I was running into issues where the zsh config would get set _after_
other configs. This was super annoying because it meant weird
workarounds when doing things like tweaking the path. By moving the zsh
config up front (and changing the general structure of the zsh config),
it's easier to reason about this kind of thing.