tmux: Add number of INBOX emails to tmux status
I like to keep an eye on how many emails I have in my various inboxes.
The best thing for me to do is to count the number of files in all of my
inboxes and print them in my tmux status bar.

If there is 1 email, this will print "1 email". Multiple emails will be
pluralized ("42 emails"). No emails will not print anything.