mutt: Update email-count to be a little smarter
Instead of just hard-coding a delimiter and ignoring zero values, this
will be a little more flexible if we make those options. This adds a
`-z` flag to `email-count` for ignoring zero values, and a `-d` flag for
setting the trailing delimiter.

The result here is that our tmux output doesn't have a delimiter unless
`email-count` is printing text.