vim: Create spelling dictionary on vim startup
Adding words to the `spellfile` isn't actually enough to add them to the
dictionary. You also need to create the dictionary from those words by
running `mkspell`. That creates a binary file that Vim reads for
spell checking.

However, we really don't want to add that file to the git repo, because
it's fairly opaque. Instead, it makes more sense (IMO) to simply
recreate that file from our custom `spellfile` every time we launch Vim.
This lets us add and remove words easily, while ensuring that we always
have the most up to date version of the spell check dictionary.