mutt: Consolidate mbsync and notmuch into mutt tag
Pulling these out into their own tag was overkill. These are
dependencies of mutt, and they don't work properly without each other.
Instead of pretending that they are decoupled, we should just accept the
fact that this is a intermingled web of tools that is being held
together by duct tape and bailing wire.