setup: Only look for launch agents in specific dir
Looking for any plist files was really naive. We have since added an
Xcode template file, and that would get picked up by this script and
treated as a launch agent script. We should avoid that.

Since we're being strict with our folder structure anyway, we can rely
on the fact that our launch agents will be located in
tag-*/source/LaunchAgents, and just look there for any plists. If I put
a plist in there that I don't want treated as a launch agent, that's my
fault.