vim: Use a plugin to pull in our color scheme
Instead of sourcing all these different color schemes in (of which we
used one), we can instead use a plugin to pull in the colorscheme we
want to use. In this case, Lucius.

Turns out, however, that you don't have _access_ to the color scheme
till after we call `vundle#end()`. That means that we can't just drop
the plugin definition in at the top of `colorscheme.vim`, because vim
will complain about not having a colorscheme called Lucius.

The fix is to ditch `colorscheme.vim` and instead to just set the
colorscheme in `vimrc` directly after loading our plugins.

Bummer.