vim: search quietly and open a quickfix
By default, `:grep` prints the search results to the console and then
brings you to the first result. This is less than ideal, since we'd have
to open the quickfix window ourselves.

To solve this, we can do a few things:

1) set up an autocommand to call `cwindow` every time we run grep. That
   will ensure that if there were search results, we have a list of them
   open.
2) use `:grep!` instead of `:grep`. This will perform the search, but it
   _won't_ jump to the first result.
3) run `:grep!` silently. This will prevent it from jumping out of vim.

Unfortunately, we can't configure `:grep` to do 2 and 3 by default, so
we need to set up a custom command for it. I've chosen `Gr` to mimic the
default command of `gr`.