zsh: don't source zsh config twice
So I'm an idiot, and of _course_ `~/.config/*` is going to match
`~/.config/zsh`. The result of this is that (among other things that I
don't know about I'm sure), our path was being blown out, since the path
defined by zsh was set last.

To fix, we can enable `EXTENDED_GLOB`, which is _awesome_ and allows us
to do things like say `~/.config/^zsh/*.zsh` which will match every file
in `~/.config/*/.zsh` _except_ files in `~/.config/zsh`. So that's
awesome, and exactly what we're looking for.

Note that this means `^` becomes a special character, to I had to quote
`^R` in our bindkey override.