git: add git-fuzzy-add using pick
We can use [pick] along with git-ls-files to give us a fuzzy-finder for
git-add.

git-ls-files will normally show all files. The --modified flag will show
us only changed files (won't show staged files). The --others flag will
show us _all_ untracked files in the directory. The --exclude-standard
flag will make sure that of our untracked files, we don't see anything
that we're already trying to ignore with git.

Passing those to pick will let us do a fuzzy search for the file we
want.

Then we can pass the chosen file to git-add.

Note that we are also passing any flags in to git-add. This is so that
you can do something like git fuzzy-add -p and only add parts of a file.

[pick]: https://github.com/thoughtbot/pick