vim: Re-enable set hidden
This lets us background buffers that have unwritten changes.

Pro: This means we can move around easier. We don't need to worry about
writing every time we want to switch buffers (like when flipping back
and forth between a couple of buffers)

Con: This makes it _really_ easy to end up with a bunch of changes that
only exist in memory. That can cause issues if Vim crashes, or if we're
trying to build something without realizing that we haven't written a
bunch of changes to disk.