rcm: Use rcm to link files into ~/Library
We can use `UNDOTTED` to symlink files into our `Library` folder,
instead of needing to symlink these manually. This is _super_ helpful,
and lets us maintain the same file structure we're used to.

Also rename the "Swift File" Xcode template to "Swift". The bit about it
being a file is pretty well implied.

Note that this doesn't _currently_ work because "File Templates" has a
space in it. This is super annoying, but should be fixed soon.