mutt: Display message dates in local time zone
This has been bugging me _forever_.

By default, Mutt just shows you the date of the sent message as
specified by the outgoing server. That means it's formatted in UTC or in
some random person's local timezone. This isn't super useful info.

To fix this, we can use `formail` to pull the date out of the original
message, use `gdate` (GNU `date`) to reformat that date into the local
timezone, then `formail` to re-assign that date back to the message.
Then we continue with our normal display filtering.

Yes, I realize this is insane.