git: Prevent myself from running git gc in gitsh
When not in gitsh, `gs` prints the git status.

However, when we _are_ in gitsh, `gs` autocompletes to `gc` which is
short for garbage collect and so I have the really bad habit of garbage
collecting my git repos multiple times a day.

I am not smart.

To prevent myself from doing this, I can just set up an alias for `gs`
so that it doesn't autocorrect.