zsh: Don't verify history
With `HIST_VERIFY` set, commands like `!!` won't actually _execute_ the
previous command, but will instead auto-fill a prompt with the previous
command, which you then have to execute manually. This didn't really
bother me, but it means that sending `!!` to a tmux split from Vim
doesn't actually do what I want it to since I then have to switch over
to that split and execute the command.

Additionally, I run these commands so infrequently that it's not that
big of a deal.