tmux: Improve display of session name
Previously, we weren't setting the status-left-length variable, so we were
getting our session name clipped. This was annoying as hell. Setting the
length to 60 (which is the same as the right side width) helps to keep this
from happening.

Additionally, it looks nice to have the session name wrapped in square
brackets.