mutt: Remap half-page scrolling to work like vim
C-d is mapped to `delete-thread` by default, which _always_ bites me. I'd
rather use a consistent keybinding for moving around text, which means copying
the vim keybindings of C-d to move half a page down and C-u to move half a
page up.

This also means we need to re-map ghu to C-x, which feels a little nicer
anyway.