git: Use diff-highlight as a diff filter
The diff-highlight script that ships as a part of git-contrib does some fancy
shit with git diffs, like highlighting the exact words that changed in a line
instead of just highlighting the whole thing. Super fancy.

I've wrapped the diff-highlight script in a local script in order to avoid
needing to muck around with my path.