git: Enable new compactionHeuristic option
This means that if you change something like:

```ruby
def foo
  bar
end
```

to something like:

```ruby
def foo
  bar
end

def baz
  bar
end
```

The diff will be highlighted properly to show the added line and the new
method, as opposed to highlighting everything between `def foo` and the final
`end`.

Note that this is how this is _supposed_ to work but I haven't been able to
get it to do the right _or_ wrong thing consistently. Whatever. Lets give it a
go.