vim: Replace custom search with Ag.vim
This grep setup was acting wonky for me and wasn't perfect. It failed on
matches and I was never sure if it was an actual failure or if it was because
of the way I set it up. Too much annoyance for me. Instead, by using Ag.vim,
everything just works the way I expect.