haskell: Point to custom openssl location
New versions of macOS don't ship with the openSSL headers... or they aren't
exposed... or something like that. I'm not entirely sure. All I know is that I
can't build HsOpenSSL because of this.

The fix is to tell Stack about these additional include and lib directories
for our custom openssl install (via homebrew)