mutt: Use followup folders for emails
Instead of keeping emails sitting in my inbox, it's better to track the email
using a system like OmniFocus and put the email in a central location so that
I can easily find it when I need it. This means using account-specific
Followup folders and re-using our ^T shortcut for adding emails to OmniFocus.