mutt: Switch to mu for searching
notmuch is pretty bulky and requires a lot of additional config that I'd
rather not deal with. Additionally, @keith came up with a pretty slick idea
for just searching the _current_ mailbox, instead of all mailboxes. I really
like that because it means we can avoid awkwardly sending emails from the
wrong account after searching.