tmux: Forward navigation keys to more programs
fzf uses C-j and C-k to move up/down in the list. However, because tmux
is grabbing those keystrokes it's causing havok. fzf.vim uses a tmux
split to display/filter the results, which means that hitting C-k will
send us back up to the vim pane when it's actually waiting for input
from fzf. So then I need to kill fzf with C-c, but that leaves the
orphaned fzf instance in the tmux pane. Annoying.

To fix this, I can refactor the way I'm deciding to forward these
keystrokes a bit and add fzf itself, as well as bash. I need to add bash
because fzf.vim wraps the raw command in a simple bash script, and so is
reported as such. AFAIK, I'm not sitting in an open bash session any
other time, so this _should_ be safe. We'll see if it causes issues down
the line.