mutt: Update search macros for an empty index
Using `<enter-command>` means that I can't use it when the index is
empty. This really sucks and I'd really like to be able to search
regardless of if the index has any emails in it or not.

I'm not 100% sure this totally fixes it, but at least the macro works
now.