vim: Accept zero arguments for Remove
More often than not, I want to remove the file I'm currently in. This
means I was constantly typing `:Remove %<CR>`, which was tedious and
annoying.

To fix this, I can modify the command so that it takes 0 or 1 arguments,
and then default to the current file path if no arguments were passed
in. But we can still pass an arbitrary file name to the command and have
it remove that file.