tmux: Let me jump to a session's specific window
Jumping to a session is cool, but it's _really_ cool to be able to jump
to a specific window _inside_ a session. Turns out, `switch-client` lets
you do that by passing the session and window name to the `-t` flag
separated by a colon.

So then the only other thing we need to do is list our windows in
addition to our sessions and append the window name to the session name.
Then we can pass that directly into `fzf` and pass _that_ directly into
`tmux switch-client`.

Dope.