tmux: Weight most recently used window to the top
_most_ of the time if I'm switching to a session what I'd really like to
do is switch to the last window that I was in for a session. Currently,
the windows are sorted by their physical position in the session, and so
that might not be the first result when searching for a session.

To fix this, we can use the included `window_active` format to show help
us perform a light-weight weighted sort on the windows so that whichever
window is currently active is the one that we list first. We can then
pass the `--no-sort` flag to `fzf` to make sure that it presents the
results in the order that we gave them, as opposed to its default
behavior of sorting them during search.