tmux: Show sessions in fuzzy-choose-tree
Turns out, if two windows have the same name, using `switch-client -t`
will fail silently. This means that if a session has two windows and
they are both named the same, it would be impossible to jump to it the
way this used to work.

To fix this, we can make sure to include the session name in the list.

As a side effect of this change, we can actually back away from our
previous solution of weighting the results. Instead, selecting the
session name itself will perform the same function.