tmux: Don't list windows if there is only one
Currently, when we list sessions and windows, we're listing all of the
windows in every session. This leads to some redundancy if there is only
one window in the session. For example, if a session `foo` has a single
window `bar`, the current behavior is to list the session and then the
session:window like so:

```
foo
foo:bar
```

Both of these refer to the same window, and so it doesn't actually
matter which we choose. This means that all we end up doing is add noise
to the list.

The better behavior would be to list windows for the session _only_ when
there is more than one window. That would mean that the previous
situation would result in:

```
foo
```

Adding another window `baz` to that session would then cause
fuzzy-choose-tree to list the windows, and so we'd end up with:

```
foo
foo:bar
foo:baz
```