mutt: use flowed text
So flowed text is a format for writing emails where I add whitespace to
the end of my hard-wrapped lines which indicates that you can re-wrap
those lines on your end as needed. Mutt supports this with a little help
from vim.

However, it doesn't seem like anyone else actually cares about this so
I'm not convinced this will have much effect. It feels nice to be able
to be morally superior about Doing The Right Thing though, so it's
worth it.