gpg: use pinentry-mac
So this is _super_ annoying but the only way I could get git to
cooperate with gpg 2.1 was to switch my pinentry over to pinentry-mac.
So now, instead of staying in my curses interface I need to deal with a
native popup window (and an ugly one at that).

But, hey. At least I'm able to sign commits again!