vim: Add bindings for re-opening files in splits
I _constantly_ open a file in the current split when I meant to open it
in a new horizontal or vertical split. These keybindings let me correct
that mistake by re-opening the current file in a split and then jumping
back to the file that was open before I screwed up in the first place.