xcode: Don't hardcode xcode name
We want to be a bit more flexible with our xcode naming, so it makes
more sense for us to dynamically look up the name of the currently
selected Xcode version, as opposed to hard-coding it as 'Xcode'.

This also gives me the oportunity to turn the `play` aliases into
functions so that we can re-use this logic.