vim: Default to neovim
I want to try out neovim for a while and see how I feel about it. Thankfully,
the config for getting neovim set up based on an existing vim config is pretty
minimal. The only two real changes I need to make are to only use the
vim-sensible plugin in vim (these are the same defaults as in neovim), and
specify the plugin directory for vim-plug. Not doing this results in vim-plug
not being able to find the existing plugins when running under neovim.