git: Let create-pull-request know about templates
If the current project has a PULL_REQUEST_TEMPLATE, we should use it!
This looks for local and hidden PR templates and adds them to the
message if they exist.

Also taking this time to rename some stuff and remove the auto-insert
stuff, which kept tripping me up.

As with most things, idea and partial implementation stolen from @gabebw