zsh: Add note function
This is stolen largely from @thorncp, and he's made multiple improvements
since I last looked at this:

https://github.com/thorncp/dotfiles/commit/b44f36003a6dfa0823caf7560549869e806933b7

The intent here is to try and improve my note taking habits by making it
easier to take notes. It's been hit or miss for me recently, I just need to
work on developing a better habit of it.