tmux: Set gitsh pane to 83 characters
Setting this as a percentage resulted in an inconsistent experience opening
these panes. I think I'll enjoy this more if it always opens to a specific
width that I know will work well for git. In this case, I'll base it off the
accepted width of git commit messages and make it 83 columns wide. This will
make the pane wide enough for the 78 characters of a git commit message body,
plus the width of the Vim gutter/line numbers.