tmux: Extract auto-start functionality into script
This moves what was a function that was automatically sourced into a command
that is explicitly run. Besides the simplicity gained by not having this be an
implicit thing that happens to be run at an indeterminate time during launch,
it's nice to have this as a command I can run from elsewhere if I choose.