meta: Consolidate more zsh config
I'm still keeping some separate files for now, but this moves a bunch of the
smaller individual files that were implicitly run into either the main zshrc
or into files that are explicitly sourced. This should help with some clarity
around when things are executed, and simplifies our setup a bit.