tmux: Consolidate configuration in tmux.conf
As with vim and zsh, the idea here is to reduce the complexity of our dotfiles
setup and move things into standard locations. This should help keep a better
mental model of everything that is happening for these tools, and should also
reduce the complexity for people trying to use our setup as an example.