git: Simplify git hooks
We had this super complicated system (that I'm not convinced actually worked)
to regenerate our ctags after specific events. We were using this system
because we wanted to be able to add tag-specific hooks and have them
integrated cleanly. Now that we've moved away from that tag structure, that
doesn't really make sense anymore and we can just use these hooks directly.