ruby: Switch to asdf for managing ruby versions
asdf is a language-agnostic version manager, which is super neat. This means
we can use it to install/manage a whole bunch of different languages instead
of needing to rely on language-specific tools. They have a very nice Ruby
plugin that we can use instead of needing chruby.

We're actually already using asdf for managing elixir and erlang, so this
commit also adds some setup for making sure we can get back to this state
easily.