tmux: Show battery info in the status bar
This is a _super_ simple little script to tell us the battery percentage and
(if unplugged) approximately how much time we have left on the battery.

Since I'm spending more and more time in tmux, this seems important. I've
found myself showing the menu bar a lot in order to get at this info.