zsh: Move NOTES env var back into my main config
I moved this out of here a while ago, but it's probably best to move it back.
This is going to be used from a few places, so it's best to keep it centrally
located.

Also update the comments to be worded in the... first person? I think? Idk
grammar.