scripts: Convert note into a script
Previously, this was implemented as a set of functions. That worked fine when
I was sourcing function files into my zshrc, but I'm not doing that anymore. I
also dislike the ways that using a function differs from using a script, such
as the behavior with cd.

So I'm going to move this into a full-fledged executable instead. This meant
(learning and) writing a zsh completion script as well, which was a fun
exercise.

One thing to... note here is that I'm also modifying the behavior of
`note new` so that it dumps the file into an `_inbox` context if you don't
provide a context. I'm also going to have it create a new file every time the
command is run, as opposed to appending to a file for today. This fits in a
bit better with my personal workflow (or, rather, the workflow I'd like to
have).