zsh: Set location env vars in zshenv
I've never been 100% sure what to put in your zshenv or why, but I learned
something tonight!

If you launch a temporary pane from tmux (a pane that is launched with a
specific command and that closes when the command is done), your zshenv is
still sourced, while your zshrc isn't. This means that if you try to run a
script that relies on environment variables (like `note new`, for example) in
a temporary tmux pane, those variables won't be set and unusual things will
happen.

We can fix this by setting these locations in the zshenv, which _is_ sourced
by those temporary panes.

In general, I think it makes more sense for stuff like this (and, possibly, the
PATH (or at least part of the path?)) set way up front like this. It's super
light weight to set these and it's nice to make sure they are available
everywhere.