notes: Make the context directory much earlier
It makes more sense to me to make this directory immediately after getting the
context, as opposed to waiting till just before making the file.

I guess the assumption here might have been that delaying this creation means
that if we misspell the context we have a chance to back out before creating a
directory on disk? I'm not sure I'm super worried about that.