tmux: Add keybinding for quickly creating a note
This is neat. Now I can hit `C-s n` (previously bound to next pane, which I
never use) and quickly open a temp tmux pane with a new note. This should give
me a nice way to quickly dump a note into the inbox without needing to move to
non-vim panes or anything like that.