zsh: Remove zsh-syntax-highlighting
This just updated and broke a _staggering_ amount of stuff for me. Most of it
was around unicode support, but it also started screwing with the layout in
mutt. It took me a while to figure out what was actually going on. I'd rather
not rely on something that can cause that much undue chaos in my life.