mutt: Don't indent macros
Indenting these macros _usually_ seemed to work, but this was actually sending
a literal tab to mutt when executed. The end result was that running these
commands from an empty index would fail because it would try to sync the
mailbox (mapped to Tab), and there would be nothing to sync.

By removing the indentation, the tab isn't sent and so these macros can be run
from an empty index.