setup: specify dotfiles dir during setup
Turns out: if you try to use an ENV var that you haven't set yet in your
current shell, it doesn't exist. This is important because we're setting
DOTFILES in our zshenv, which hasn't been sourced when we run our setup. So,
we actually need to specify to use the current directory as the dotfiles dir.