git: Wrap hub browse to open BitBucket Server urls
This replaces our old functionality of `hub browse` with a custom `git-browse`
script that is able to open BitBucket Server urls as well as GitHub ones. We
can do this by checking the origin url and matching it against some known
patterns. This relies on some private config not shared in this repo
(specifically, the stash url pattern env var).