git: Don't default to using a graph
Using this as the default isn't great, since the project I'm in most often has
a pretty messy graph representation. What I _really_ want is to see a nice
clean line, and that's easiest to get to by not using `--graph`. Omitting this
here will also let me add it when I actually want it.