git: Reformat git log
Now that I'm on a team with more people, I found this format harder to read.
Moving the name/gpg info to the beginning and the refs to the end makes it
much easier to read the actual commit messages, since the left side is now a
fairly consistent width.