git: colocate git-update
This has been kept in an external repo up till now, but that almost certainly
isn't required. I'm not positive anyone other than me is actually using this
anymore, and it's so incredibly trivial at this point that maintaining a
homebrew tap for it is just an unnecessary amount of overhead. Instead, by
bringing it into a local script, I can just modify it to behave the way I like
without needing to make it generic enough for other people.