git: Always use color = auto for git-wtf
I really just want this to be auto all the time. I don't really ever want to
change this. Auto is actually the default for color.ui as well, so this makes
it match.

I should _probably_ make this default to `auto` instead of just forcing it to
`auto` but who cares.