git: Automatically delete merged branches
If a branch that we have locally checked out has been merged into master, we
want to delete our local copy of that branch. This is useful for me right now
because 99% of my branches are merged via web interface, as opposed to merging
it locally. Automating this branch deletion process (which was pulled from my
existing git-delete-local-merged command) saves me from having to remember to
run this additional command manually.