vim: Add simple plugin for opening jira tickets
It's not uncommon to run across jira tickets in various projects. This simple
command/function opens the ticket under the specific cursor, using a
separately defined root domain. This domain should be of the format
'https://jira.company.com/browse/'.

This plugin is super naive right now, but gets the job done.