git: Add wip/resume commands
My old habit of hammering `git wip` and just walking away worked well if I
came back to the commit quickly, but caused _major_ issues when I walked away
for a long enough period of time to forget what I was doing. At that point,
failing tests or failing builds aren't enough to get me back on track.

What I'd _like_ to do, is get in the habit of taking notes around what I was
trying to accomplish when I paused, and this script is an attempt to help with
that. `git wip` will now launch vim and prompt for me to enter notes for what
I was working on. Then, when I'm ready to continue, `git resume` will print
those notes out before resetting that commit.