git-wip: Use standalone buffer for notes
Instead of editing directly in the commit message, it's nice to be able to
enter the notes in isolation, so that I don't even need to care about the
commit subject.

To do this, we'll create another temp file and open it for markdown editing.
We'll then use that as the body of the commit message directly, instead of
using it as a template.