git: Use completions from git instead of zsh
I've been using the zsh completions for a while, but the git ones are actually
better. Unfortunately, they never played well with my setup and I couldn't
ever figure out why.

Looks like it was because of a combination of a couple of things:

* `setopt complete_aliases`: I don't know what I thought this did or why I
needed it but it seems like it was preventing aliases from being used for
completion.
* My git aliases themselves: because I was using `g` instead of `git`, the
completion engine wasn't able to figure out what to use for completion. I
think. I'm still not 100% sure on _either_ of these tbh but this seems to work
the way I expect now so that's good.

This _does_ break completion for my custom `git delete-branch` subcommand, and
I'm working on figuring out how to get that work properly. But at least
there's a lot of wackyness removed from my setup now.