scripts: Add Dropbox Uploader script
GitHub: https://github.com/andreafabrizi/Dropbox-Uploader

This script uploads arbitrary files to Dropbox, which is crazy useful to be
able to do on machines that don't have Dropbox.app installed. This is the case
with my work computer, but having this script allows me to keep Dropbox
archival as a core part of my workflow for things like web-based receipts.