git: fuzzy-checkout and make it the default action
Instead of having to remember branch names all the damn time I'd rather search
for them. This currently only looks at local branches, but could be expanded
to remotes in the future I suppose. It also skips the fzf searching if you
pass in an argument, which is good because I have strong muscle memory for
things like `gco master` and `gco -`.