vim: use vim-polyglot for language support
This plugin replaces a few of the language plugins I was using, and adds
support for like 200 more. It also does the thing I never did properly and
only loads them when needed, so it should be faster than including individual
plugins for languages.

kudos to @drapergeek for letting me know this existed