tmux: Tweak the appearance of window tabs
This adds a gray background around my current window, and flips the color
palette to be black on yellow if that window is currently zoomed.

I'd _love_ to find a way to combine the two `window_zoomed_flag` conditionals,
but I can't figure out how to escape the comma between `fg` and `bg`.