tmux: add new command for fuzzy-choose-tree
This used to be mapped to <leader>p, but we'll re-add it as <leader>P because
the behavior of jumping to a specific window was kind of useful and it might
be nice to see a list of all open sessions sometimes. We'll see how much I end
up using this in reality.