vim: reload files on focus
A _constant_ thing I run into is that I'll be working on something in
vim, flip over to Xcode to do something (run a test, build, whatever)
and then end up making a small edit. I then immediately forget about
that edit and go back to working in vim, where I then immediately forget
about the changes I made in Xcode, only to be reminded when I try to
save.

We can fix this though! Tmux has the ability to send focus events, and
then we can add a trigger to run `:checktime` on the current buffer when
we get those events (or when we enter the buffer). This works really
well for me! I guess I might get hit with issues if I have unsaved
changes in vim, but I haven't run into that case yet, so we'll see!