python: Use asdf to manage python versions
I'm _constantly_ fighting with the process of installing python
correctly via homebrew, and it doesn't have to be this way. Since I'm
already all-in with asdf for version management, it makes way more sense
to simply use asdf to manage python as well. There's one step here where
we set the global python versions that I'm not doing here, but this
should still be a improved experience overall.