tmux: Replace extract_url with script
extract_url breaks for me fairly often, which is obnoxious. Replacing it
with a shell script seems like a completely reasonable thing to do. I'm
stealing this one from wxfr/tmux-fzf-url, which normally works as a tmux
plugin but I've avoided TPM thus far and don't see the need to install
it now. I've also customized the script a _bit_ to fit my needs better.