haskell: set prompt-cont
Without this set, the default prompt would be use for multiline input.
That prompt would be _real weird_ because it includes all of the
currently loaded modules, so I'd jump from a nice short prompt to
suddenly getting pushed way out to the right. This new prompt is nicer
and still visually distinct from the main prompt (it uses the default
color and a different glyph), so it's easy to tell when I'm working on a
multiline input.