brew: Update homebrew paths/setup
Homebrew installs packages in a different location now, I'm not sure if it's
because I'm on an M1 chip, or because of an OS update. Either way, I needed to
do some tweaking to my setup scripts to include the homebrew path when run,
and I needed to update/remove some hard-coded executable paths.