nvim: Add LSP/Formatting/Completion config
This is a pretty large code dump but it pulls the basic
LSP/Formatting/Completion config from nvim-lua/kickstart.nvim and gets us up
and running with some basic Lua completion behavior. I'm also leaving all of
the documentation from kickstart in place as I continue to tinker and tweak
this config.