tmux: Revamp status bar with Nova
Everyone else seems to use catppuccin(1), which I do agree is "prettier". But
I could not for the life of me get it to work properly in my setup. I have to
assume it's a "me" problem but life is too short for this crap. So I'm using
Nova instead, which is a _much_ simpler plugin (with a much simpler design)
but still gets me a prettier status bar! Huzzah!

Also note that I'm moving it to the top because the triple stack of
status/command/status while in (n)vim has always bothered me and now it's
better.

1: https://github.com/catppuccin/tmux