nvim: Set foldlevelstart
I used to have foldlevel set like this but foldlevelstart is actually
the correct setting to use, because it does _not_ mess with the fold
level once the file is open it just sets the fold level for the file at
the point when it is opened. This is important because using foldlevel
alone makes commands like zm (close more) not work properly unless you
hit it 99 times (in order to pass the specified zoom level). So this solves my
problem (It drives me nuts to open a completely folded file) without breaking
folds themselves.