zsh: Export XDG_CONFIG_HOME
Tools _should_ automatically look at ~/.config, but they sometimes don't
(cough cough lazygit cough cough). To fix this we'll go ahead and export
XDG_CONFIG_HOME, which is probably a good practice anyway.

I didn't know quite where to put this in the rc file so I put it with
the path because ¯\_(ツ)_/¯