tmux: Specify config directory for lazygit
I love lazygit but it sure does need a lot of handholding for
configuration on macos. Or I guess for using XDG_CONFIG_HOME properly.
Probably not their fault but it's obnoxious.

Anyway, since lazygit is the default command for this window it seems to
_not_ be getting the XDG_CONFIG_HOME env variable properly, so I need to
explicitly point it in the right place.