lazygit: Set default branch prefix
I basically always name my branches like this (especially at work), so
it makes a lot of sense to have this default. The nice thing is that
since this is being set via the lazygit prompt I can chose to clear out
the prefix if I don't want to use it in this one specific case.