fish: Add pcd command
This gives me access to the crux of the tat script (quickly jumping into a
project directory) without the rest of the tmux session cruft. I'm likely
going to ditch tmux soon in favor of native splits, so this will help me to
keep some of the important behavior I need from that workflow.