nvim: Disable auto-formatting for Swift code
This is unfortunate, but my work codebase is _wildly_ out of sync with
the autoformatting preference and so using this feature results in more
headaches than are worth it. In the future I might try to find a way to
conditionally re-enable this for specific codebases, but I'll wait till
I need that behavior to do that work.