vim: add new split management keybindings
I've gone through a number of iterations on these and this feels the
best so far.

New splits are created with Ctrl+\ and Ctrl+-. This is a very old binding that
I lost at some point, but I really like the pneumonic of - for horizontal
splits and \ (or |) for vertical splits.

For moving between splits, I can't use Ctrl+<Arrow> because that's bound to
various window management stuff in macOS. But Opt+<Arrow> is unbound at the
system level, and as a bonus the commands for moving between Ghostty splits is
Cmd+Opt+<Arrow> so there's a nice symmetry there. However, for key input
reasons I don't fully understand, the right and left arrows come through as
"f" and "b" when paired with Opt. So we're mapping those instead of using
<Left> and <Right>. Side-note: while in insert mode, Ctrl+k will let you hit a
key (or a chord) and have it print out the digraph for it. This was
_incredibly_ useful for understanding why these keybindings weren't working
like I expected them to.

I probably need some resizing bindings in here but I'll get to those later on.